+++ email_routing/rule_test.go
@@ -29,11 +29,11 @@ func TestRuleNewWithOptionalParams(t *testing.T) {
 	)
 	_, err := client.EmailRouting.Rules.New(context.TODO(), email_routing.RuleNewParams{
 		ZoneID: cloudflare.F("023e105f4ecef8ad9ca31a8372d0c353"),
-		Actions: cloudflare.F([]email_routing.ActionParam{{
+		Actions: cloudflare.F([]email_routing.ActionParam{email_routing.ActionParam{
 			Type:  cloudflare.F(email_routing.ActionTypeForward),
 			Value: cloudflare.F([]string{"destinationaddress@example.net"}),
 		}}),
-		Matchers: cloudflare.F([]email_routing.MatcherParam{{
+		Matchers: cloudflare.F([]email_routing.MatcherParam{email_routing.MatcherParam{
 			Type:  cloudflare.F(email_routing.MatcherTypeLiteral),
 			Field: cloudflare.F(email_routing.MatcherFieldTo),
 			Value: cloudflare.F("test@example.com"),
@@ -71,11 +71,11 @@ func TestRuleUpdateWithOptionalParams(t *testing.T) {
 		"a7e6fb77503c41d8a7f3113c6918f10c",
 		email_routing.RuleUpdateParams{
 			ZoneID: cloudflare.F("023e105f4ecef8ad9ca31a8372d0c353"),
-			Actions: cloudflare.F([]email_routing.ActionParam{{
+			Actions: cloudflare.F([]email_routing.ActionParam{email_routing.ActionParam{
 				Type:  cloudflare.F(email_routing.ActionTypeForward),
 				Value: cloudflare.F([]string{"destinationaddress@example.net"}),
 			}}),
-			Matchers: cloudflare.F([]email_routing.MatcherParam{{
+			Matchers: cloudflare.F([]email_routing.MatcherParam{email_routing.MatcherParam{
 				Type:  cloudflare.F(email_routing.MatcherTypeLiteral),
 				Field: cloudflare.F(email_routing.MatcherFieldTo),
 				Value: cloudflare.F("test@example.com"),
@@ -87,6 +87,34 @@ func TestRuleUpdateWithOptionalParams(t *testing.T) {
 			Source:         cloudflare.F(email_routing.RuleUpdateParamsSourceAPI),
 		},
 	)
+	if err != nil {
+		var apierr *cloudflare.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestRuleListWithOptionalParams(t *testing.T) {
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := cloudflare.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIKey("144c9defac04969c7bfad8efaa8ea194"),
+		option.WithAPIEmail("user@example.com"),
+	)
+	_, err := client.EmailRouting.Rules.List(context.TODO(), email_routing.RuleListParams{
+		AccountID: cloudflare.F("account_id"),
+		Enabled:   cloudflare.F(email_routing.RuleListParamsEnabledTrue),
+		Page:      cloudflare.F(1.000000),
+		PerPage:   cloudflare.F(5.000000),
+	})
 	if err != nil {
 		var apierr *cloudflare.Error
 		if errors.As(err, &apierr) {